+++ scripts/ac5e-settings.mjs
@@ -24,6 +24,7 @@ export default class Settings {
 	static AUTOMATE_STATUSES = 'automateStatuses';
 	static REMOVE_NON5E_STATUSES = 'displayOnly5eStatuses';
 	static CONTEXT_KEYWORDS_REGISTRY = 'contextKeywordsRegistry';
+	static USAGE_RULES_REGISTRY = 'usageRulesRegistry';
 	static CONTEXT_KEYWORDS_ALLOW_PLAYER_PERSIST = 'contextKeywordsAllowPlayerPersist';
 	static DEV_MODE_ENABLED = 'devModeEnabled';
 
@@ -226,6 +227,13 @@ export default class Settings {
 			default: {},
 			type: Object,
 		});
+		game.settings.register(Constants.MODULE_ID, Settings.USAGE_RULES_REGISTRY, {
+			name: 'Usage rules registry',
+			scope: 'world',
+			config: false,
+			default: {},
+			type: Object,
+		});
 		game.settings.register(Constants.MODULE_ID, Settings.CONTEXT_KEYWORDS_ALLOW_PLAYER_PERSIST, {
 			name: 'Allow player context keyword persistence',
 			scope: 'world',
@@ -317,5 +325,8 @@ export default class Settings {
 	}
 	get contextKeywordsAllowPlayerPersist() {
 		return game.settings.get(Constants.MODULE_ID, Settings.CONTEXT_KEYWORDS_ALLOW_PLAYER_PERSIST);
+	}
+	get usageRulesRegistry() {
+		return game.settings.get(Constants.MODULE_ID, Settings.USAGE_RULES_REGISTRY);
 	}
 }